Gater: create unit test with assertions
Created a unit test that will actually test functionality automatically,
without needing to look through waveforms. It will check each time step
is as it should be, and will fail if the gateware is not correct.

Found a bug in gateware_utils.MockPhy using this, where its output
signal was too far delayed.

This is to provide a ground-truth for whether the InputGater (renamed
TriggeredInputGater) is functioning properly for future modifications.