Gaters: reset timestamp on clear
Timestamp would remain when cleared, which doesn't tell you if the
gater has actually registered an event. The downside to this is that
the Gater timestamp will be overwritten as soon as the Entangler is
started the next time, but that's not too big of a deal, I hope.
If it is, revert this.