packaging: fix settings refs for package install
Couldn't find settings.toml when installing as a package,
so use pkg_resources instead. Relevant for using as a Nix package.
Will need to convert all versions to this. It should work with
``pip install -e .`` as well, I hope.

Includes text files in the distribution for the package working properly.